Add files via upload
This is the in progress code for neutral current interactions on O-16. The dominant neutral current interactions in the energy range of interest are those which result in a nucleon knockout (Suzuki et al. 2018). As such there are two code files, for neutron and proton knock-out (o16nc_n_code and o16nc_p_code respectively).
A spline function is used to produce a fit 'function' for cross-section as a function of neutrino energy. The data this spline uses is taken from Suzuki et al. 2018 with additional estimated data points (estimated from by-eye fits) to ensure a smooth spline fit. The threshold energies for the interactions have also been estimated from a by-eye fit as no theoretical or experimental research could be found into the exact values.
If the neutrino energy is large enough, the remnant nucleus (after nucleon knock-out) has ~70% chance of being left in an excited state (T2K 2019 paper). In this case two potentially detectable particles are produced, the knocked-out nucleon and a de-excitation photon.
A function, random(), generates a random number between 0 and 1 which is used to determine whether a de-excitation photon is produced. This has been written in the o16nc_n_code and will be added to the o16nc_p_code. I will probably adapt this function to a gamma(eNu) function which will return true if a de-excitation photon is produced.
A function to generate the kinematics of the knocked-out nucleon has been added to each code, however it has not been written in full yet.